+++ projects/js-packages/number-formatters/tests/number-format.test.ts
@@ -1,15 +1,7 @@
-import { beforeEach, describe, expect, it, jest } from '@jest/globals';
+import { describe, expect, it } from '@jest/globals';
 import { numberFormat, numberFormatCompact } from '../src/number-format.js';
 
-type IntlType = typeof Intl & {
-	NumberFormat: typeof Intl.NumberFormat;
-};
-
 describe( 'numberFormat()', () => {
-	beforeEach( () => {
-		jest.clearAllMocks();
-	} );
-
 	it( 'should format number with default options', () => {
 		const result = numberFormat( { browserSafeLocale: 'en-US' } ).format( 1234.56 );
 		expect( result ).toBe( '1,235' );
@@ -34,20 +26,6 @@ describe( 'numberFormat()', () => {
 		} ).format( 1234.56 );
 		expect( result ).toBe( '$1,235' );
 	} );
-
-	it( 'should return original number as string if formatting fails', () => {
-		// mock `Intl.NumberFormat` to throw an error
-		const originalIntlNumberFormat = Intl.NumberFormat;
-		jest.spyOn( Intl as IntlType, 'NumberFormat' ).mockImplementation( () => {
-			throw new Error( 'Invalid locale' );
-		} );
-
-		const result = numberFormat( { browserSafeLocale: 'en' } ).format( 1234.56 );
-		expect( result ).toBe( '1234.56' );
-
-		// restore original `Intl.NumberFormat`
-		( Intl as IntlType ).NumberFormat = originalIntlNumberFormat;
-	} );
 } );
 
 describe( 'numberFormatCompact()', () => {
@@ -70,18 +48,5 @@ describe( 'numberFormatCompact()', () => {
 			forceLatin: false,
 		} ).format( 1234567 );
 		expect( result ).toBe( '١٫٢ مليون' );
-	} );
-
-	it( 'should return original number as string if formatting fails', () => {
-		// mock `Intl.NumberFormat` to throw an error
-		const originalIntlNumberFormat = Intl.NumberFormat;
-		jest.spyOn( Intl as IntlType, 'NumberFormat' ).mockImplementation( () => {
-			throw new Error( 'Invalid locale' );
-		} );
-		const result = numberFormatCompact( { browserSafeLocale: 'invalid-locale' } ).format( 1234567 );
-		expect( result ).toBe( '1234567' );
-
-		// restore original `Intl.NumberFormat`
-		( Intl as IntlType ).NumberFormat = originalIntlNumberFormat;
 	} );
 } );
